+++ src/Elders.Cronus/MessageProcessing/MessageProcessor.cs
@@ -121,7 +121,7 @@ namespace Elders.Cronus.MessageProcessing
                             feedResult = feedResult.AppendError(message, new FeedError()
                             {
                                 Origin = new ErrorOrigin(handlerSubscription.Id, ErrorOriginType.MessageHandler),
-                                Error = new Protoreg.ProtoregSerializableException(ex)
+                                Error = new SerializableException(ex)
                             });
                         }
                     }